+++ AU/Plugins/Gist.ps1
@@ -46,7 +46,12 @@ Get-ChildItem $Path | ForEach-Object {
 # request
 
 #https://github.com/majkinetor/au/issues/142
-[System.Net.ServicePointManager]::SecurityProtocol = 3072 -bor 768 -bor [System.Net.SecurityProtocolType]::Tls -bor [System.Net.SecurityProtocolType]::Ssl3
+if ($PSVersionTable.PSVersion.major -ge 6) {
+    $AvailableTls = [enum]::GetValues('Net.SecurityProtocolType') | Where-Object { $_ -ge 'Tls' } # PowerShell 6+ does not support SSL3, so use TLS minimum
+    $AvailableTls.ForEach({[Net.ServicePointManager]::SecurityProtocol = [Net.ServicePointManager]::SecurityProtocol -bor $_})
+} else {
+    [System.Net.ServicePointManager]::SecurityProtocol = 3072 -bor 768 -bor [System.Net.SecurityProtocolType]::Tls -bor [System.Net.SecurityProtocolType]::Ssl3
+}
 
 $params = @{
     ContentType = 'application/json'
@@ -64,5 +69,7 @@ if ($ApiKey) {
 $Response = Invoke-WebRequest @params
 if ($Response.StatusCode -in @(200, 201, 304)) {
     $JsonResponse = $Response.Content | ConvertFrom-Json
-    Write-Output $JsonResponse.html_url
+    $GistURL = $JsonResponse.html_url
+    $Revision = $JsonResponse.history[0].version
+    Write-Output "$GistURL/$Revision"
 }