+++ AU/Plugins/Report/text.ps1
@@ -4,10 +4,10 @@ $Title       = if ($Params.Title) { $Params.Title } else {  'Update-AUPackages' 
 #==============================================================================
 
 function title($txt) { "`r`n{0}`r`n{1}`r`n" -f $txt,('-'*$txt.Length) }
-function indent($txt, $level=4) { $txt -split "`n" | % { ' '*$level + $_ } }
+function indent($txt, $level=4) { $txt -split "`n" | ForEach-Object { ' '*$level + $_ } }
 
 $now         = $Info.startTime.ToUniversalTime().ToString('yyyy-MM-dd HH:mm')
-$au_version  = gmo au -ListAvailable | % Version | select -First 1 | % { "$_" }
+$au_version  = Get-Module au -ListAvailable | ForEach-Object Version | Select-Object -First 1 | ForEach-Object { "$_" }
 $package_no  = $Info.result.all.Length
 
 "{0,-15}{1}" -f 'Title:', $Title
@@ -28,33 +28,33 @@ else {
 
 if ($Info.pushed) {
     title Pushed
-    $Info.result.pushed | select 'Name', 'Updated', 'Pushed', 'RemoteVersion', 'NuspecVersion' | ft | Out-String | set r
+    $Info.result.pushed | Select-Object 'Name', 'Updated', 'Pushed', 'RemoteVersion', 'NuspecVersion' | Format-Table | Out-String | Set-Variable r
     indent $r 2
 
-    $Info.result.pushed | % { $_.Name; indent $_.Result; "" }
+    $Info.result.pushed | ForEach-Object { $_.Name; indent $_.Result; "" }
 }
 
 if ($Info.error_count.total) {
     title Errors
-    $Info.result.errors | select 'Name', 'NuspecVersion', 'Error' | ft | Out-String | set r
+    $Info.result.errors | Select-Object 'Name', 'NuspecVersion', 'Error' | Format-Table | Out-String | Set-Variable r
     indent $r 2
 
-    $Info.result.errors | % { $_.Name; indent $_.Error; "" }
+    $Info.result.errors | ForEach-Object { $_.Name; indent $_.Error; "" }
 }
 
 
 if ($Info.result.ignored) {
     title Ignored
-    $Info.result.ignored | ft | select 'Name', 'NuspecVersion', 'IgnoreMessage' | ft | Out-String | set r
+    $Info.result.ignored | Format-Table | Select-Object 'Name', 'NuspecVersion', 'IgnoreMessage' | Format-Table | Out-String | Set-Variable r
     indent $r 2
 }
 
-$ok = $Info.result.ok | ? { !$_.Pushed }
+$ok = $Info.result.ok | Where-Object { !$_.Pushed }
 if ($ok) {
     title OK
-    $ok | select 'Name', 'Updated', 'RemoteVersion', 'NuspecVersion' | ft | Out-String | set r
+    $ok | Select-Object 'Name', 'Updated', 'RemoteVersion', 'NuspecVersion' | Format-Table | Out-String | Set-Variable r
     indent $r 2
 
-    $ok | % { $_.Name; indent $_.Result; "" }
+    $ok | ForEach-Object { $_.Name; indent $_.Result; "" }
 }