+++ AU/Public/Push-Package.ps1
@@ -13,19 +13,22 @@ function Push-Package() {
     param(
         [switch] $All
     )
-    $api_key =  if (Test-Path api_key) { gc api_key }
-                elseif (Test-Path ..\api_key) { gc ..\api_key }
+    $api_key =  if (Test-Path api_key) { Get-Content api_key }
+                elseif (Test-Path (Join-Path '..' 'api_key')) { Get-Content (Join-Path '..' 'api_key') }
                 elseif ($Env:api_key) { $Env:api_key }
 
     $push_url =  if ($Env:au_PushUrl) { $Env:au_PushUrl }
                  else { 'https://push.chocolatey.org' }
+                 
+    $force_push = if ($Env:au_ForcePush) { '--force' }
+                  else { '' }
 
-    $packages = ls *.nupkg | sort -Property CreationTime -Descending
-    if (!$All) { $packages = $packages | select -First 1 }
+    $packages = Get-ChildItem *.nupkg | Sort-Object -Property CreationTime -Descending
+    if (!$All) { $packages = $packages | Select-Object -First 1 }
     if (!$packages) { throw 'There is no nupkg file in the directory'}
     if ($api_key) {
-        $packages | % { cpush $_.Name --api-key $api_key --source $push_url }
+        $packages | ForEach-Object { choco push $_.Name --api-key $api_key --source $push_url $force_push }
     } else {
-        $packages | % { cpush $_.Name --source $push_url }
+        $packages | ForEach-Object { choco push $_.Name --source $push_url $force_push }
     }
 }