+++ build.ps1
@@ -27,7 +27,7 @@ param(
 $b = {
     if ($Clean) { git clean -Xfd -e vars.ps1; return }
     if ($ShortVersion) { $Version = [string] $Version = [Version](Get-Date).ToUniversalTime().ToString("yyyy.M.d") }
-    if ($LastCommitDate) { $Version = [string] $Version = [Version]$(git log -1 --date=short)[3].split(' ')[-1].replace("-",".") }
+    if ($LastCommitDate) { $Version = [string] $Version = [Version]$(((git log -1 --date=short) | Where-Object { $_ -match "date:"}) | Select-Object -First 1).split(' ')[-1].replace("-",".") }
 
     $module_path    = "$PSScriptRoot/AU"
     $module_name    = Split-Path -Leaf $module_path
@@ -44,7 +44,7 @@ $b = {
     create_manifest
     create_help
 
-    cp $installer_path $build_dir
+    Copy-Item $installer_path $build_dir
     zip_module
     build_chocolatey_package
 
@@ -58,32 +58,32 @@ function zip_module() {
 
     $zip_path = "$build_dir\${module_name}_$version.7z"
     $cmd = "$Env:ChocolateyInstall/tools/7z.exe a '$zip_path' '$module_path' '$installer_path'"
-    $cmd | iex | Out-Null
+    $cmd | Invoke-Expression | Out-Null
     if (!(Test-Path $zip_path)) { throw "Failed to build 7z package" }
 }
 
 function init() {
     if ($remove_old) {
         Write-Host "Removing older builds"
-        rm -Recurse (Split-Path $build_dir) -ea ignore
+        Remove-Item -Recurse (Split-Path $build_dir) -ea ignore
     }
-    mkdir -Force $build_dir | Out-Null
-    cp -Recurse $module_path $build_dir
+    New-Item -Type Directory -Force $build_dir | Out-Null
+    Copy-Item -Recurse $module_path $build_dir
 }
 
 function build_chocolatey_package {
     if ($NoChocoPackage) { Write-Host "Skipping chocolatey package build"; return }
 
     & $PSScriptRoot/chocolatey/build-package.ps1
-    mv "$PSScriptRoot/chocolatey/${module_name}.$version.nupkg" $build_dir
+    Move-Item "$PSScriptRoot/chocolatey/${module_name}.$version.nupkg" $build_dir
 }
 
 function create_help() {
     Write-Host 'Creating module help'
 
     $help_dir = "$module_path/en-US"
-    mkdir -Force $help_dir | Out-Null
-    gc $PSScriptRoot/README.md | select -Skip 4 | Set-Content "$help_dir/about_${module_name}.help.txt" -Encoding ascii
+    New-Item -Type Directory -Force $help_dir | Out-Null
+    Get-Content $PSScriptRoot/README.md | Select-Object -Skip 4 | Set-Content "$help_dir/about_${module_name}.help.txt" -Encoding ascii
 }
 
 function create_manifest() {