+++ tests/test_package_with_streams/tools/chocolateyInstall.ps1
@@ -1,7 +1,7 @@
 $ErrorActionPreference = 'Stop'
 
 $packageName = 'test_package_with_streams'
-$url32       = gcm choco.exe | % Source
+$url32       = Get-Command choco.exe | ForEach-Object Source
 $checksum32  = ''
 
 $params = @{